+++ inferno-core/inferno-core.cabal
@@ -35,6 +35,7 @@ library
     , Inferno.Parse
     , Inferno.Parse.Commented
     , Inferno.Parse.Error
+    , Inferno.Utils.Arbitrary
     , Inferno.Utils.QQ.Script
     , Inferno.Utils.QQ.Module
   other-modules:
@@ -59,6 +60,7 @@ library
     , parser-combinators   >= 1.3.0 && < 1.4
     , picosat              >= 0.1.6 && < 0.2
     , prettyprinter        >= 1.7.1 && < 1.8
+    , QuickCheck           >= 2.14.2 && < 2.15
     , recursion-schemes    >= 5.2.2.3 && < 5.3
     , syb                  >= 0.7.2 && < 0.8
     , template-haskell     >= 2.15 && < 2.19
@@ -102,6 +104,7 @@ test-suite inferno-tests
     , prettyprinter
     , pretty-simple
     , QuickCheck
+    , quickcheck-instances
     , recursion-schemes
     , text
   default-language:    Haskell2010
