+++ inferno-vc/inferno-vc.cabal
@@ -46,6 +46,8 @@ library
     , http-client                        >= 0.7.13 && < 0.8
     , http-types                         >= 0.12.3 && < 0.13
     , inferno-types                      >= 0.1.0 && < 0.2
+    -- TODO why doesn't this work?
+    -- , inferno-core
     , lens                               >= 5.2 && < 5.3
     , mtl                                >= 2.2.2 && < 2.3
     , plow-log                           >= 0.1.6 && < 0.2
